refactor: Rename scripts
