+++ src/scripts/livekit.ts
@@ -1,339 +1,74 @@
-// @ts-nocheck
-import puppeteer from 'puppeteer-core';
-import fs from 'fs';
-import {logMessage, asProcessArg} from '../utils';
-import {LogLevel} from '../types/global';
-import {getToken} from "./token";
+import {dynamicImport} from "tsimportlib";
 
 // Variables
-let browser: any = null;
-let page: any = null;
-let ready: boolean = false;
+let token: string;
+let tokenCreatedAt: number;
 
 /**
- * Setter to change ready value
+ * Generate a token for LiveKit (valid for 6 hours)
  *
- * @param {boolean} value - New value
- * @returns {void}
+ * @returns {Promise<string>} Token
  */
-function setReady(value: boolean): void {
-    ready = value;
-}
-
-/**
- * Value of the environnement variable
- *
- * @param name
- * @returns {string | undefined} Value of the variable
- */
-function getEnvVariable(name: string): string | undefined {
-    return process.env[name];
-}
-
-/**
- * Show / Send data to the parent process
- *
- * @param {string} message - Message to send
- * @param {LogLevel} type - Type of message
- * @returns {void}
- */
-function reportMessage(message: string, type: LogLevel = LogLevel.INFO): void {
-    if (process.send) {
-        process.send(message);
-    } else {
-        logMessage(message, type);
-    }
-}
-
-/**
- * Launch headless browser
- *
- * @returns {Promise<any>} Instance of the browser
- */
-export async function newPage(): Promise<any> {
-    if (browser) {
-        if (!page) {
-            page = await browser.newPage();
-        }
-        return page;
+export async function getToken(): Promise<string> {
+    if (token && tokenCreatedAt && (Date.now() - tokenCreatedAt) < 60 * 60 * 6 * 1000) {
+        return token;
     }
 
-    // Prepare args
-    const args: string[] = [
-        '--disable-setuid-sandbox',
-        '--no-sandbox',
-        '--enable-gpu',
-        '--use-fake-ui-for-media-stream',
-        '--autoplay-policy=no-user-gesture-required'
-    ]
-    if (asProcessArg('fake-devices')) {
-        args.push('--use-fake-device-for-media-stream')
+    // Import SDK
+    const liveKitSDK: any = await dynamicImport('livekit-server-sdk', module) as typeof import('livekit-server-sdk');
+
+    // Check room
+    const roomService = new liveKitSDK.RoomServiceClient(
+        process.env['LIVEKIT_HTTP_URL'],
+        process.env['LIVEKIT_KEY'],
+        process.env['LIVEKIT_SECRET']
+    );
+
+    const rooms: any = await roomService.listRooms();
+    const createRoom: boolean = !rooms.some((room: any): any => {
+        if (room.name === process.env['LIVEKIT_ROOM']) {
+            return true;
+        }
+    })
+
+    console.log(createRoom);
+    if (createRoom) {
+          console.log(await roomService.createRoom({
+            emptyTimeout: 999999999,
+            name: process.env['LIVEKIT_ROOM']
+          }));
     }
 
-    // Launch the browser
-    browser = await puppeteer.launch({
-        executablePath: '/usr/bin/chromium',
-        args: args,
-        ignoreDefaultArgs: [
-            '--mute-audio',
-            '--hide-scrollbars'
-        ]
+    // Generate a new token
+    const accessToken: any = new liveKitSDK.AccessToken(
+        process.env['LIVEKIT_KEY'],
+        process.env['LIVEKIT_SECRET'],
+        {
+            identity: "Car",
+            ttl: 60 * 60 * 7,
+        },
+    );
+
+    // Set permissions
+    accessToken.addGrant({
+        roomCreate: false,
+        roomJoin: true,
+        roomList: false,
+        roomRecord: false,
+        roomAdmin: false,
+        room: process.env['LIVEKIT_ROOM'],
+        ingressAdmin: false,
+        canPublish: true,
+        canSubscribe: true,
+        canPublishData: true,
+        canUpdateOwnMetadata: true,
+        hidden: false,
+        recorder: false,
+        agent: false
     });
 
-    // Allow permissions
-    const context: any = browser.defaultBrowserContext();
-    await context.overridePermissions(process.env['LIVEKIT_HTTP_URL'], ['microphone', 'camera']);
+    token = await accessToken.toJwt();
+    tokenCreatedAt = Date.now();
 
-    // Open new page
-    page = await browser.newPage();
-    return page;
+    return token;
 }
-
-/**
- * Launch stream
- *
- * @returns {Promise<void>}
- */
-export async function startStream(): Promise<void> {
-    const page = await newPage();
-
-    page.on('console', async (msg: any): Promise<void> => {
-        const msgArgs = msg.args();
-        for (let i = 0; i < msgArgs.length; ++i) {
-            reportMessage(await msgArgs[i].jsonValue());
-        }
-    });
-
-    page.on('pageerror', async (error: string): Promise<void> => {
-        reportMessage(error, LogLevel.ERROR);
-    });
-
-    await page.goto(process.env['LIVEKIT_HTTP_URL'], {waitUntil: 'load'});
-
-    await page.exposeFunction('getToken', getToken);
-    await page.exposeFunction('setReady', value => setReady(value));
-    await page.exposeFunction('asArg', argument => asProcessArg(argument));
-    await page.exposeFunction('getEnvVariable', name => getEnvVariable(name));
-
-    const script: string = fs.readFileSync(`${__dirname}/../libs/livekit-client.min.js`, 'utf8');
-    await page.addScriptTag({content: script});
-
-    await page.evaluate(async (): Promise<void> => {
-        await window.setReady(true)
-
-        // Variables
-        let room = null;
-        const defaultDevices = await asArg('default-devices');
-        const fakeDevices = await asArg('fake-devices');
-        const noCamera = await asArg('no-camera');
-        const tracks = {
-            video: null,
-            audio: null
-        };
-
-        // Send data to room
-        async function dataEvent(event) {
-            if (!room) {
-                return;
-            }
-
-            await room.localParticipant.publishData(
-                new TextEncoder().encode(JSON.stringify(event.detail.data)),
-                LivekitClient.DataPacket_Kind.LOSSY
-            );
-        }
-
-        // Create local video and audio tracks
-        async function createTracks() {
-            const devices = await navigator.mediaDevices.enumerateDevices();
-            const videoDevices = devices.filter(device => device.kind === 'videoinput');
-            const audioDevices = devices.filter(device => device.kind === 'audioinput');
-
-            const createVideoTrack = async (deviceId, width = 640, height = 360, maxBitrate = 400_000, priority = 'high') => {
-                if (!deviceId) {
-                    return null;
-                }
-
-                return await LivekitClient.createLocalVideoTrack({
-                    deviceId: deviceId,
-                    resolution: {
-                        width: width,
-                        height: height,
-                        encoding: {
-                            maxFramerate: 24,
-                            maxBitrate: maxBitrate,
-                            priority: priority,
-                        }
-                    }
-                });
-            }
-
-            const createAudioTrack = async (deviceId, channelCount = 1) => {
-                if (!deviceId) {
-                    return null;
-                }
-
-                return await LivekitClient.createLocalAudioTrack({
-                    deviceId: deviceId,
-                    autoGainControl: true,
-                    echoCancellation: false,
-                    noiseSuppression: false,
-                    channelCount: channelCount
-                });
-            }
-
-            if (videoDevices.length > 0 && !noCamera) {
-                tracks.video = await createVideoTrack(
-                    videoDevices.find(device => device.label.startsWith('Cam Link 4K'))?.deviceId
-                );
-            }
-
-            if (audioDevices.length > 0) {
-                tracks.audio = await createAudioTrack(
-                    audioDevices.find(device => device.label.startsWith('Cam Link 4K'))?.deviceId
-                );
-            }
-
-            try {
-                console.log('Connecting to LiveKit...');
-                setTimeout(startSession);
-            } catch {
-                console.log('Crash. Restarting...');
-                window.removeEventListener('data', dataEvent)
-                room = null
-                setTimeout(startSession);
-            }
-        }
-
-        // Connect to LiveKit room and publish tracks with datas
-        async function startSession() {
-            const token = await window.getToken();
-            const url = await window.getEnvVariable('LIVEKIT_WS_URL');
-
-            room = new LivekitClient.Room({
-                adaptiveStream: true,
-                dynacast: true,
-                reconnectPolicy: {
-                    nextRetryDelayInMs: (context) => {
-                        return 600;
-                    }
-                }
-            });
-
-            await room.prepareConnection(url, token);
-
-            room
-                .on(LivekitClient.RoomEvent.Connected, () => {
-                    console.log('Connected');
-                    window.addEventListener('data', dataEvent);
-                })
-                .on(LivekitClient.RoomEvent.Reconnecting, () => {
-                    console.log('Reconnecting...');
-                    window.removeEventListener('data', dataEvent);
-                })
-                .on(LivekitClient.RoomEvent.Reconnected, () => {
-                    console.log('Reconnected');
-                    window.addEventListener('data', dataEvent);
-                })
-                .on(LivekitClient.RoomEvent.Disconnected, () => {
-                    console.log('Disconnected. Restarting...');
-                    window.removeEventListener('data', dataEvent);
-                    room = null
-                    setTimeout(startSession);
-                });
-
-            await room.connect(url, token);
-
-            if (fakeDevices || defaultDevices) {
-                const devices = await navigator.mediaDevices.enumerateDevices();
-
-                if (devices.filter(device => device.kind === 'videoinput').length > 0 && !noCamera) {
-                    await room.localParticipant.setCameraEnabled(true);
-                }
-
-                if (devices.filter(device => device.kind === 'audioinput').length > 0) {
-                    await room.localParticipant.setMicrophoneEnabled(true);
-                }
-            } else {
-                const publishVideoTrack = async (track, stream = 'main', maxBitrate = 400_000, priority = 'high') => {
-                    if (!track) {
-                        return;
-                    }
-
-                    await room.localParticipant.publishTrack(track, {
-                        name: `${stream}-video`,
-                        stream: stream,
-                        source: 'camera',
-                        simulcast: false,
-                        videoCodec: 'AV1',
-                        videoEncoding: {
-                            maxFramerate: 24,
-                            maxBitrate: maxBitrate,
-                            priority: priority
-                        }
-                    });
-                }
-
-                const publishAudioTrack = async (track, stream = 'main', maxBitrate = 12_000) => {
-                    if (!track) {
-                        return;
-                    }
-
-                    await room.localParticipant.publishTrack(track, {
-                        name: `${stream}-audio`,
-                        stream: stream,
-                        source: 'audio',
-                        simulcast: false,
-                        red: true,
-                        dtx: true,
-                        stopMicTrackOnMute: false,
-                        audioPreset: {
-                            maxBitrate: maxBitrate
-                        }
-                    });
-                }
-
-                if (!noCamera && tracks.video) {
-                    await publishVideoTrack(tracks.video);
-                }
-
-                if (tracks.audio) {
-                    await publishAudioTrack(tracks.audio);
-                }
-            }
-        }
-
-        if (fakeDevices || defaultDevices) {
-            console.log('Connecting to LiveKit...');
-            try {
-                setTimeout(startSession);
-            } catch {
-                console.log('Crash. Restarting...');
-                window.removeEventListener('data', dataEvent)
-                room = null
-                setTimeout(startSession);
-            }
-        } else {
-            console.log('Creating tracks...');
-            setTimeout(createTracks);
-        }
-    });
-}
-
-// Send data to the room
-process.on('message', (data: any): void => {
-    if (browser && page && ready) {
-        page.evaluate((data: any): void => {
-            const customEvent: CustomEvent = new CustomEvent(
-                'data',
-                {
-                    detail: {
-                        data: data
-                    }
-                });
-            window.dispatchEvent(customEvent);
-        }, data);
-    }
-});
-
-setTimeout(startStream);
